Extracted ServerAppRunner and ClientAppRunner.
